update performance report layout
Pembenahan tipe data decimal 2 untuk score, dan tampilan report modal detail

+++ database/migrations/2021_07_10_024116_create_capaian_kpi_table.php
@@ -31,7 +31,7 @@ class CreateCapaianKpiTable extends Migration {
             $table->string('polarization');
             $table->string('timeframe_target');
             $table->bigInteger('capaian');
-            $table->bigInteger('score');
+            $table->double('score', 20, 2);
             $table->bigInteger('weightedscore');
             $table->timestamps();
         });
